added stock window confirm on delete actions and removed window.reload

+++ frontend/src/components/HUD.tsx
@@ -16,6 +16,8 @@ function HUD({
 
 	const handleDelete = async () => {
 		try {
+			const confirm = window.confirm('Are you sure you want to delete this cycle?');
+			if (!confirm) return;
 			await deleteCycle(user?.token!, cycleId);
 			setCycleId(null);
 		} catch (error) {
@@ -24,7 +26,9 @@ function HUD({
 	};
 	return (
 		<div className="w-[400px] z-10 bg-secondary h-[70%] border border-tertiary rounded p-2 px-4 flex items-center justify-between">
-			<h1 className="flex justify-center items-center font-bold">
+			<div className='flex'>
+
+			<h1 className="flex justify-center items-center font-bold mr-2">
 				Completion:
 			</h1>
 			<div className="completion-circle h-14">
@@ -37,13 +41,16 @@ function HUD({
 							style={{
 								strokeDasharray: 2 * 42.5 * 3.14,
 								strokeDashoffset:
-									2 * 42.5 * 3.14 * (1 - calcTotal(tasks) / 100),
+								2 * 42.5 * 3.14 * (1 - calcTotal(tasks) / 100),
 							}}
-						/>
+							/>
 					</svg>
 					<div className="inner">{calcTotal(tasks)}%</div>
 				</div>
 			</div>
+							</div>
+			<div className='flex'>
+
 			<button className="transition-all hover:brightness-125 border border-primary-content rounded ml-2 bg-accent h-8 w-8 flex items-center justify-center">
 				<svg
 					className="fill-primary-content"
@@ -51,24 +58,25 @@ function HUD({
 					viewBox="0 0 16 16"
 					width="16"
 					height="16"
-				>
+					>
 					<path d="M0 8a8 8 0 1 1 16 0A8 8 0 0 1 0 8Zm1.5 0a6.5 6.5 0 1 0 13 0 6.5 6.5 0 0 0-13 0Zm10.28-1.72-4.5 4.5a.75.75 0 0 1-1.06 0l-2-2a.751.751 0 0 1 .018-1.042.751.751 0 0 1 1.042-.018l1.47 1.47 3.97-3.97a.751.751 0 0 1 1.042.018.751.751 0 0 1 .018 1.042Z"></path>
 				</svg>
 			</button>
 			<button
 				onClick={handleDelete}
 				className="transition-all hover:brightness-125 border border-primary-content rounded ml-2 bg-error h-8 w-8 flex items-center justify-center"
-			>
+				>
 				<svg
 					className="fill-primary-content"
 					xmlns="http://www.w3.org/2000/svg"
 					viewBox="0 0 12 12"
 					width="12"
 					height="12"
-				>
+					>
 					<path d="M2.22 2.22a.749.749 0 0 1 1.06 0L6 4.939 8.72 2.22a.749.749 0 1 1 1.06 1.06L7.061 6 9.78 8.72a.749.749 0 1 1-1.06 1.06L6 7.061 3.28 9.78a.749.749 0 1 1-1.06-1.06L4.939 6 2.22 3.28a.749.749 0 0 1 0-1.06Z"></path>
 				</svg>
 			</button>
+					</div>
 		</div>
 	);
 }
